add view binding to rate and rateAfter fragments
