+++ bandyer-android-design/src/main/res/layout/fragment_rate_after.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RateAfterFragment">
+    tools:context=".rating.RateAfterFragment">
 
 </FrameLayout>
